move page block code from App.jsx to Header.jsx
